pyproject.toml
[project]
name = "mcp-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.14.3",
    "langchain>=1.2.6",
    "langchain-google-genai>=4.2.0",
    "langchain-mcp-adapters>=0.2.1",
    "python-dotenv>=1.2.1",
]
